% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert data table(s) to data frame(s) with rownames}
\usage{
convert_to_df(dt_or_lst)
}
\arguments{
\item{dt_or_lst}{A data table or (possibly list of) list(s) of data tables
where each dt has a unique ID in its first column.}
}
\value{
A corresponding data frame or (possibly list of) list(s) of data
frames, in which the unique ID column has been transformed into row names.
}
\description{
Convert data table(s) into data frame(s), also transforming the first 
(assumed ID) field into row names. Can be used to e.g. convert a single 
\code{\link[=sum_by_clade]{summed}} data table, or all data table elements in
\code{\link[=load_data]{loaded}} or \code{\link[=merge_data]{merged}} lists.
}
\details{
Converts data table(s) into data frame(s), also transforming the
first (assumed ID) field into row names. If an ID is missing, when e.g.
an ASV lacks a name at some taxonomic rank, it will be given row name
'Unknown'.

Example usage:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{merged <- merge_data(loaded)}}{}
  \item{\code{merged_df <- convert_to_df(merged)}}{}
}
}
