% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert data table(s) to data frame(s) with rownames}
\usage{
convert_to_df(dt_obj)
}
\arguments{
\item{dt_obj}{A data table or (possibly hierarchical) list of data tables
where each dt has a unique ID in its first column.}
}
\value{
A corresponding data frame or (possibly list of) list(s) of data
frames, in which the unique ID column has been transformed into row names.
}
\description{
Convert data table(s) into data frame(s), also transforming the first 
(assumed ID) field into row names. Can be used to e.g. convert a single 
\code{\link[=sum_by_clade]{summed}} data table, or all data table elements in
\code{\link[=load_data]{loaded}} or \code{\link[=merge_data]{merged}} lists.
}
\details{
Converts data table(s) into data frame(s), also transforming the
first (assumed ID) field into row names. ASVs that lack an ID, i.e. a name
at some taxonomic rank, will be grouped into a row named 'Unnamed-clades' at
that level.

Example usage:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{merged <- merge_data(loaded)}}{}
  \item{\code{merged_df <- convert_to_df(merged)}}{}
}
}
