% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{load_data}
\alias{load_data}
\title{Load downloaded ASV occurrence data}
\usage{
load_data(data_path = './datasets');
}
\arguments{
\item{data_path}{Path of directory containing dataset (*.zip) files}
}
\value{
A list of four sublists (\code{counts}, \code{asvs}, \code{events},
  \code{emof}) containing sparse matrix or data table elements from each 
  dataset, indexed by \code{datasetID}.
}
\description{
Load Amplicon Sequence Variant (ASV) occurrence data from 'Darwin
Core (DwC)-like' archives downloaded from the Swedish ASP portal,
\url{https://asv-portal.biodiversitydata.se/}.
}
\details{
Reads data from one or more compressed archives. Returns a list of
  sub lists, each of which contains sparse matrix or data table 
  objects from each included dataset:

\itemize{
  \item \strong{counts}: List of sparse matrices representing read counts
    (taxon [row] x event [col] sparse matrix) from each dataset.
    \itemize{
      \item \code{`first-datasetID`} (sparse matrix)
      \item \code{`second-datasetID`} (sparse matrix)
      \item ...
    }

  \item \strong{asvs}: List of data tables containing the DNA sequence and
    taxonomic assignment of ASVs (taxon [row] x attribute [col]) from each dataset.

    \itemize{
      \item \code{`first-datasetID`} (data table)
      \item \code{`second-datasetID`} (data table)
      \item ...
    }

  \item \strong{events}: List of data tables representing basic event/sample
    metadata (event [row] x parameter [col]) from each dataset.

    \itemize{
      \item \code{`first-datasetID`} (data table)
      \item \code{`second-datasetID`} (data table)
      \item ...
    }

  \item \strong{emof}: List of data tables representing additional contextual
    parameter values (event [row] x measurementType [col]) from each dataset.
    \itemize{
      \item \code{`first-datasetID`} (data table)
      \item \code{`second-datasetID`} (data table)
      \item ...
    }
}

To inspect individual matrices or data tables:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{View(loaded$emof$`first-datasetID`)}}{}
  \item{\code{# OR (to show first 100 ASVs in first counts matrix):}}{}
  \item{\strong{Memory note:} `counts` is a sparse matrix; converting large
    matrices to dense format (e.g. `as.matrix()`) may exhaust RAM. Only convert
    small subsets for inspection.}{}
  \item{\code{View(as.matrix(loaded$counts[[1]][1:100,]))}}{}
}
}
