% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{load_data}
\alias{load_data}
\title{Load downloaded ASV occurrence data}
\usage{
load_data(data_path = './datasets');
}
\arguments{
\item{data_path}{Path of directory containing dataset (*.zip) files}
}
\value{
A list of four sublists (\code{counts}, \code{asvs}, \code{events},
  \code{emof}) containing sparse matrice or data table elements from each 
  dataset, indexed by \code{datasetID}.
}
\description{
Load Amplicon Sequence Variant (ASV) occurrence data from 'Darwin
Core (DwC)-like' archives downloaded from the Swedish ASP portal,
\url{https://asv-portal.biodiversitydata.se/}.
}
\details{
Reads data from one or more compressed archives. Returns a list of
  sub lists, each of which contains sparse matrice or data table 
  objects from each included dataset:

\itemize{
  \item \strong{counts}: List of sparse matrices representing read counts
    (\code{taxonID} [row] x \code{eventID} [col] matrix) from each dataset.
    \itemize{
      \item \code{`first-datasetID`} (sparse matrix)
      \item \code{`second-datasetID`} (sparse matrix)
      \item ... 
    }

  \item \strong{asvs}: List of data tables containing the DNA sequence and 
  taxonomic assignment of ASVs in a \code{taxonID} [row] x attribute [col] format.
    
    \itemize{
      \item \code{`first-datasetID`} (data table)
      \item \code{`second-datasetID`} (data table)
      \item ... 
    }

 \item \strong{event}: List of data tables representing basic event/sample metadata
    in a \code{eventID} [col] x parameter [row] format for each dataset.
    
    \itemize{
      \item \code{`first-datasetID`} (data table)
      \item \code{`second-datasetID`} (data table)
      \item ... 
    }

  \item \strong{emof}: List of data tables representing additional contextual
    parameter values (\code{measurementValue}) in a \code{eventID} [row]
    x \code{measurementType} [col] format for each dataset.
    \itemize{
      \item \code{`first-datasetID`} (data table)
      \item \code{`second-datasetID`} (data table)
      \item ... 
    }
}

To inspect individual matrices or data tables:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{View(loaded$emof$`first-datasetID`)}}{}
  \item{\code{# OR (to show first 100 ASVs in first counts matrix):}}{}
  \item{\code{View(as.matrix(loaded$counts[[1]][1:100,]))}}{}
}
}
