% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{load_data}
\alias{load_data}
\title{Load downloaded ASV occurrence data}
\usage{
load_data(data_path = './datasets');
}
\arguments{
\item{data_path}{Path of directory containing dataset (*.zip) files}
}
\value{
A list of four sublists (\code{counts}, \code{asvs}, \code{events},
  \code{emof}) containing data table elements from each dataset, indexed by
  \code{datasetID}.
}
\description{
Load Amplicon Sequence Variant (ASV) occurrence data from 'Darwin
Core (DwC)-like' archives downloaded from the Swedish ASP portal,
\url{https://asv-portal.biodiversitydata.se/}.
}
\details{
Reads data from one or more compressed archives. Returns a list of
  sub lists, each of which contains data table objects (dt:s) from each
  included dataset:

\itemize{
  \item \strong{counts}: List of dt:s representing read counts
    (\code{taxonID} [row] x \code{eventID} [col] matrix) from each dataset.
    \itemize{
      \item \code{`first-datasetID`} (dt)
      \item \code{`second-datasetID`} (dt)
    }

  \item \strong{asvs}: List of dt:s representing `asv_sequence` and taxonomy
    per \code{taxonID} from each dataset.
    \itemize{
      \item ...
    }

 \item \strong{event}: List of dt:s representing basic event/sample metadata
    in a \code{eventID} [col] x parameter [row] matrix from each dataset.
    \itemize{
      \item ...
    }

  \item \strong{emof}: List of dt:s representing additional contextual
    parameter values (\code{measurementValue}) in a \code{eventID} [row]
    x \code{measurementType} [col] matrix from each dataset.
    \itemize{
      \item ...
    }
}

To access individual dataset tables:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{View(loaded$counts$`first-datasetID`)}}{}
  \item{\code{# OR:}}{}
  \item{\code{View(loaded$counts[[1]])}}{}
}
}
