% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{load_data}
\alias{load_data}
\title{Read and reshape downloaded ASV occurrence data}
\usage{
load_data(data_path = './datasets');
}
\arguments{
\item{data_path}{Path of directory containing dataset (*.zip) files}
}
\value{
A list of three sublists (\code{asv_tables}, \code{asvs},
\code{emof_tables}) containing data.table elements from each dataset, 
indexed by \code{datasetID}.
}
\description{
Read and reshape Amplicon Sequence Variant (ASV) occurrence data from
'Darwin Core (DwC)-like' archives downloaded from the Swedish ASP portal, 
\url{https://asv-portal.biodiversitydata.se/}.
}
\details{
The function reads text files (\code{occurrence/asv/emof.tsv}) from compressed
archives (\code{[datasetID].zip}), and repackages data into three 
data.table objects per dataset:

\enumerate{
  \item ASV-table: Read counts in a \code{taxonID} [row] x \code{eventID} [col] matrix.
  \item asvs: \code{asv_sequence}, and taxonomy columns per \code{taxonID}.
  \item EMOF-table: Contextual parameter values (\code{measurementValue}) in a
    \code{measurementType} (\code{measurementUnit}) x \code{eventID} matrix.
}

Tables from different datasets are indexed by their respective \code{datasetID},
and organized into three lists (\code{asv_tables}, \code{asvs}, \code{emof_tables}).
These lists are returned as elements of a parent list.

To access individual dataset tables:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{View(loaded$asv_tables$`[datasetID]`)}}{ # OR:}
  \item{\code{View(loaded$asvs[[1]])}}{}
}
}
