% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{merge_data}
\alias{merge_data}
\title{Merge data from different ASV occurrence datasets}
\usage{
merge_data(loaded);
}
\arguments{
\item{loaded}{A list of three sublists (\code{counts}, \code{asvs},
\code{emof}) from \code{load_data()} function.}
}
\value{
A list of three data.table elements (\code{counts}, \code{asvs},
\code{emof}) containing data merged from loaded datasets.
}
\description{
Merge data from different datasets previously loaded with
\code{load_data()} function.
}
\details{
Takes the output from \code{load_data()} and merges rows from
different ASV occurrence datasets into three data.table objects:

\enumerate{
  \item counts: Read counts in a \code{taxonID} [row] x \code{eventID} [col] matrix.
  \item asvs: \code{asv_sequence}, and taxonomy columns per \code{taxonID}.
  \item emof: Contextual parameter values (\code{measurementValue}) in a
    \code{measurementType} (\code{measurementUnit}) x \code{eventID} matrix.
}

Merged counts and emof tables include the UNION of unique rows
(i.e. ASVs or measurements) and events (i.e. sequenced samples)
from loaded datasets. Table asvs includes non-dataset-specific data, only, 
so that we get a single row per unique ASV. Resulting tables are
returned as elements of list.

To access an individual table:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{merged <- merge_data(loaded)}}{}
  \item{\code{View(merged$counts)}}{}
}
}
