% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_processing.R
\name{sum_by_clade}
\alias{sum_by_clade}
\title{Sum counts by clade within taxonomic ranks}
\usage{
sum_by_clade(counts, asvs)
}
\arguments{
\item{counts}{A data.table with ASV read counts in a taxonID [row] x eventID
[col] matrix, from a \code{\link[=load_data]{loaded}} or
\code{\link[=merge_data]{merged}} ASV occurrence dataset.}

\item{asvs}{A data.table containing the taxonomic classification of ASV:s
included in the \code{counts} data.table.}
}
\value{
A list containing two sub-lists: `raw` and `norm`, each including
  data tables for summed ASV counts at each taxonomic rank.
}
\description{
Sum raw and normalized ASV counts by distinct clades at different taxonomic
ranks, for each sample in a \code{\link[=load_data]{loaded}} or
\code{\link[=merge_data]{merged}} ASV occurrence dataset.
}
\details{
Sums raw and normalized read counts across ASVs within distinct
  clades, at specified taxonomic ranks, to provide higher-level views of the
  data. The function normalizes ASV read counts by total counts per sample,
  and returns a list of two sub list, each of which contains data tables
  (dt:s) of summed counts for each taxonomic rank:

\itemize{
  \item \strong{raw}: List of dt:s showing raw read counts summed by 
  clade at different taxonomic ranks.
    \itemize{
      \item \code{kingdom} (dt)
      \item ...
      \item \code{species} (dt)
    }
  \item \strong{norm}: List of dt:s representing normalized read counts 
  summed by clade at different taxonomic ranks.
    \itemize{
      \item \code{kingdom} (dt)
      \item ...
      \item \code{species} (dt)
    }
}
#' To view an individual table:
\describe{
  \item{\code{loaded <- load_data(data_path = './datasets')}}{}
  \item{\code{merged <- merge_data(loaded)}}{}
  \item{\code{summed <- sum_by_clade(merged$counts, merged$asvs)}}{}
  \item{\code{View(summed$raw$family)}}{}
}
}
